requirements.txt
torch==2.0.1
torchvision==0.15.2
mlflow==2.7.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
opencv-python==4.8.1.78
Pillow==10.0.0
python-dotenv==1.0.0
pyyaml==6.0
tqdm==4.65.0
tensorboard==2.13.0
black==23.7.0
flake8==6.0.0
isort==5.12.0
pytest==7.4.0
pytest-cov==4.1.0
pre-commit==3.3.3
dvc==3.50.2  # 修复：使用存在的版本
dagshub==0.3.10